refactor(_skeleton): share state interface
