+++ packages/components/src/components/_skeleton/internal/functional-components/skeleton/component.tsx
@@ -1,6 +1,6 @@
 import type { FunctionalComponent as FC } from '@stencil/core';
 import { h } from '@stencil/core';
-import type { SkeletonState } from './controller';
+import type { SkeletonState } from './state';
 
 export const SkeletonFC: FC<SkeletonState> = ({ nameState, showState }) => {
 	return showState && <span>{nameState}</span>;
