+++ packages/components/src/components/_skeleton/internal/functional-components/skeleton/state.ts
@@ -1,0 +1,7 @@
+import type { NamePropType } from './schema/props/name';
+import type { ShowPropType } from './schema/props/show';
+
+export interface SkeletonState {
+	nameState?: NamePropType;
+	showState?: ShowPropType;
+}
